Переезд сорудников в отдельный контроллер
