+++ WebStore/Views/Employees/Index.cshtml
@@ -1,0 +1,38 @@
+﻿@model IEnumerable<WebStore.Models.Employee>
+
+    <div class="container">
+        <h3>Сотрудники</h3>
+
+        <hr />
+
+        <table class="table">
+            <thead>
+                <tr>
+                    <th>Id</th>
+                    <th>Фамилия</th>
+                    <th>Имя</th>
+                    <th>Отчество</th>
+                    <th>Возраст</th>
+                </tr>
+            </thead>
+            <tbody>
+                @foreach (var employee in Model)
+                {
+                    <tr>
+                        <td>@employee.Id</td>
+                        <td>@employee.SurName</td>
+                        <td>@employee.Name</td>
+                        <td>@employee.Patronymic</td>
+                        <td>@employee.Age</td>
+                        <td><a asp-controller="Employees" asp-action="Details" asp-route-id="@employee.Id">Информация</a></td>
+                    </tr>
+                }
+            </tbody>
+        </table>
+
+        <hr />
+        <a href="~/">На главную</a>
+        <hr />
+        <p>Тестовое значение: @ViewData["TestValue"]</p>
+
+    </div>
