--- WebStore/Views/Home/EmployeeInfo.cshtml
@@ -1,45 +1,0 @@
-﻿@model WebStore.Models.Employee
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>WebStore - Сотрудник:@Model.FullName</title>
-    <style>
-        th, td {
-            padding: 5px;
-            text-align: left;
-        }
-    </style>
-</head>
-<body>
-
-    <h3>Сотрудник</h3>
-    <table>
-        <tr class="">
-            <th>Id</th>
-            <td>@Model.Id</td>
-        </tr>
-        <tr>
-            <th>Фамилия</th>
-            <td>@Model.SurName</td>
-        </tr>
-        <tr>
-            <th>Имя</th>
-            <td>@Model.Name</td>
-        </tr>
-        <tr>
-            <th>Отчество</th>
-            <td>@Model.Patronymic</td>
-        </tr>
-        <tr>
-            <th>Возраст</th>
-            <td>@Model.Age</td>
-        </tr>
-    </table>
-
-    <hr/>
-    <a href="~/Home/Employees">К списку</a>
-</body>
-</html>
