--- WebStore/Views/Home/Employees.cshtml
@@ -1,46 +1,0 @@
-﻿@model IEnumerable<WebStore.Models.Employee>
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Employees - @ViewBag.Title</title>
-</head>
-<body>
-
-    <h3>Сотрудники</h3>
-
-    <hr />
-
-    <table>
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Фамилия</th>
-                <th>Имя</th>
-                <th>Отчество</th>
-                <th>Возраст</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var employee in Model)
-            {
-                <tr>
-                    <td>@employee.Id</td>
-                    <td>@employee.SurName</td>
-                    <td>@employee.Name</td>
-                    <td>@employee.Patronymic</td>
-                    <td>@employee.Age</td>
-                    <td><a href="~/Home/EmployeeInfo/@employee.Id">Информация</a></td>
-                </tr>
-            }
-        </tbody>
-    </table>
-
-    <hr />
-    <a href="~/">На главную</a>
-    <hr />
-    <p>Тестовое значение: @ViewData["TestValue"]</p>
-</body>
-</html>
